build: undo the splitting between "user" ans "system" bashrc
It has been causing confusion, major headaches, and an unbelievable
amount of wasted time.  The better thing to do is simply to mark
some files as "stefano-only" (i.e., to be installed only for those
user ids with which I'm operating through my usual "developer/user
persona"), and make the other files more portable and flexible so
that I can take advantage of them from more identities (e.g.,
when I'm logged in as root, or as a special system user, like
"python-master" on my Debian desktop, for example).